requirements.txt
numpy==1.18.2
pandas==1.0.2
matplotlib==3.2.0
flair==0.8.0.post1
scikit_learn==0.24.1
Keras==2.4.3
tensorflow-gpu--2.2.0





